+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.0.0 | tbd
 * Send image URL to analytics endpoint
+* Send "right-click" event when user right-clicks on search result item
 
 # v0.1.1 | 21.04.2023
 * Remove geolocation attributes from `User` object